+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/palette.module.ts
@@ -90,6 +90,8 @@ import { ReadOrganisationFieldComponent, ReadOrganisationFieldRawComponent, Read
 import { PaletteService } from './palette.service';
 import { CasePaymentHistoryViewerFieldComponent } from './payment';
 import { ReadPhoneUKFieldComponent, WritePhoneUKFieldComponent } from './phone-uk';
+import { ReadQueryManagementFieldComponent, WriteQueryManagementFieldComponent } from './query-management';
+import { QueryCreateComponent, QueryDetailComponent, QueryListComponent } from './query-management/components';
 import { ReadTextFieldComponent, WriteTextFieldComponent } from './text';
 import { ReadTextAreaFieldComponent, WriteTextAreaFieldComponent } from './text-area';
 import { UnsupportedFieldComponent } from './unsupported-field.component';
@@ -144,6 +146,7 @@ const PALETTE_COMPONENTS = [
   ReadComplexFieldCollectionTableComponent,
   ReadCaseFlagFieldComponent,
   ReadLinkedCasesFieldComponent,
+  ReadQueryManagementFieldComponent,
 
   // Write
   WriteJudicialUserFieldComponent,
@@ -162,6 +165,7 @@ const PALETTE_COMPONENTS = [
   WriteDateFieldComponent,
   WriteCaseFlagFieldComponent,
   WriteLinkedCasesFieldComponent,
+  WriteQueryManagementFieldComponent,
 
   // new
   WriteYesNoFieldComponent,
@@ -190,15 +194,7 @@ const PALETTE_COMPONENTS = [
   ReadDynamicMultiSelectListFieldComponent,
   ReadDynamicListFieldComponent,
   ReadDynamicRadioListFieldComponent,
-  LinkedCasesToTableComponent,
-  LinkedCasesFromTableComponent,
-  BeforeYouStartComponent,
-  LinkCasesComponent,
-  CheckYourAnswersComponent,
-  UnLinkCasesComponent,
-  NoLinkedCasesComponent,
-
-  // Case Flags components
+  // Components for case flags
   CaseFlagTableComponent,
   SelectFlagTypeComponent,
   SearchLanguageInterpreterComponent,
@@ -208,7 +204,19 @@ const PALETTE_COMPONENTS = [
   UpdateFlagComponent,
   CaseFlagSummaryListComponent,
   ConfirmFlagStatusComponent,
-  UpdateFlagAddTranslationFormComponent
+  UpdateFlagAddTranslationFormComponent,
+  // Components for linked cases
+  LinkedCasesToTableComponent,
+  LinkedCasesFromTableComponent,
+  BeforeYouStartComponent,
+  LinkCasesComponent,
+  CheckYourAnswersComponent,
+  UnLinkCasesComponent,
+  NoLinkedCasesComponent,
+  // Components for query management
+  QueryCreateComponent,
+  QueryDetailComponent,
+  QueryListComponent
 ];
 
 @NgModule({
@@ -242,7 +250,8 @@ const PALETTE_COMPONENTS = [
     RpxTranslationModule.forChild(),
     MatDialogModule,
     MediaViewerModule,
-    LoadingModule
+    LoadingModule,
+    RpxTranslationModule.forChild()
   ],
   declarations: [
     FixedListPipe,
@@ -275,9 +284,9 @@ const PALETTE_COMPONENTS = [
     FileUploadStateService,
     FileUploadProgressGuard,
     WindowService,
-    LinkedCasesService,
     CommonDataService,
     JurisdictionService,
+    LinkedCasesService,
     {provide: MAT_DATE_LOCALE, useValue: 'en-GB'}
   ],
   entryComponents: [ CaseFileViewFolderSelectorComponent ]