+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/query-management/domain/query-list-column/query-list-column.model.ts
@@ -1,0 +1,7 @@
+import { SortOrder } from '../../../complex/sort-order';
+
+export interface queryListColumn {
+    name: string;
+    displayName: string;
+    sortOrder: SortOrder;
+  }